Build: Keep the published lockfile and age gate in sandbox CI
The CI install step emptied yarn.lock and .yarnrc.yml and set
npmMinimalAgeGate to 0, so every run resolved the whole tree from live npm
and could install a package published hours earlier, then execute its
lifecycle scripts during sb init.

Keep both files and preapprove only the locally published Storybook
packages, which is what sandbox generation already does. Drop the
yarn set version call with them: it only worked because the config was
wiped first, and a yarnPath next to the pinned packageManager field makes
corepack abort.
