+++ scripts/sandbox/utils/yarn.ts
@@ -3,11 +3,14 @@ import { readFile, rename, rm, writeFile } from 'node:fs/promises';
 import { join } from 'path';
 
 import semver from 'semver';
-import yml from 'yaml';
 
-import { STORYBOOK_PACKAGE_PATTERNS } from '../../../code/core/src/common/js-package-manager/util.ts';
 import { ROOT_DIRECTORY } from '../../utils/constants.ts';
 import { runCommand } from '../generate.ts';
+
+export {
+  LOCALLY_PUBLISHED_PACKAGE_PATTERNS,
+  preapproveLocallyPublishedPackages,
+} from '../../utils/preapprove-local-packages.ts';
 
 interface SetupYarnOptions {
   cwd: string;
@@ -311,52 +314,6 @@ async function narrowQuarantinedRanges({
 
   // So the committed lockfile is always the product of a plain `yarn install`.
   await runCommand(`yarn install --mode=update-lockfile`, { cwd, env }, debug);
-}
-
-/**
- * Packages served by the local Verdaccio registry during sandbox generation.
- * They are published seconds before they are installed, so they can never
- * satisfy the age gate on their own.
- */
-export const LOCALLY_PUBLISHED_PACKAGE_PATTERNS = [
-  ...STORYBOOK_PACKAGE_PATTERNS,
-  'create-storybook',
-  'sb',
-];
-
-/**
- * Allow the locally published Storybook packages past the age gate for the
- * `after-storybook` install, keeping the gate itself in force.
- *
- * This install is the only step in sandbox generation that executes package
- * code (lifecycle scripts, addon postinstall hooks), and it resolves
- * third-party dependencies from the upstream registry through Verdaccio.
- * Switching the gate off here would leave the one dangerous step unprotected,
- * so name the packages that genuinely cannot satisfy it instead.
- *
- * Merges with whatever the template already allows, so a prerelease template
- * does not lose its own entries. Phase 1 strips both keys from the published
- * `after-storybook` tree, so none of this reaches consumers.
- */
-export async function preapproveLocallyPublishedPackages(cwd: string) {
-  const configPath = join(cwd, '.yarnrc.yml');
-
-  let config: Record<string, unknown> = {};
-  try {
-    config = (yml.parse(await readFile(configPath, 'utf-8')) ?? {}) as Record<string, unknown>;
-  } catch {
-    // No config yet (the lockfile refresh may have bailed); start from scratch.
-  }
-
-  const existing = Array.isArray(config.npmPreapprovedPackages)
-    ? (config.npmPreapprovedPackages as string[])
-    : [];
-
-  config.npmPreapprovedPackages = Array.from(
-    new Set([...existing, ...LOCALLY_PUBLISHED_PACKAGE_PATTERNS])
-  );
-
-  await writeFile(configPath, yml.stringify(config));
 }
 
 /**
